Remove deployment anti affinity rule.
I don't think this is required for this extension and it's somehow creating a burden on local setups because for some reason it always deploys 2 replicas even when passing 1 explicitly through the chart values. So, for single node seeds the extension will not become ready.

Only Gardener provider extensions have these constraints anyway, so I think we can just remove. It's not a problem when this extension is not available for a short time.